Use float for quantity in job status update

+++ src/utils/hact/job-status-update/update-job.js
@@ -12,7 +12,7 @@ export const buildUpdateWorkOrder = (
         customCode: task.code.split(' - ')[0],
         customName: task.description,
         quantity: {
-          amount: [Number.parseInt(task.quantity)],
+          amount: [Number.parseFloat(task.quantity)],
         },
       }
     })
